+++ Modules/SharePointDsc/DSCResources/MSFT_SPSearchServiceApp/MSFT_SPSearchServiceApp.psm1
@@ -55,6 +55,12 @@ function Get-TargetResource
 
     Write-Verbose -Message "Getting Search service application '$Name'"
 
+    if ($PSBoundParameters.ContainsKey("WindowsServiceAccount"))
+    {
+        Write-Verbose -Message ("This parameter is deprecated in this resource. Please use " + `
+                                "SPSearchServiceSettings instead.")
+    }
+
     $result = Invoke-SPDSCCommand -Credential $InstallAccount `
                                   -Arguments @($PSBoundParameters, $PSScriptRoot) `
                                   -ScriptBlock {
@@ -129,10 +135,6 @@ function Get-TargetResource
                 }
             }
 
-            $searchService = Get-SPEnterpriseSearchService
-            $windowsAccount = New-Object -TypeName System.Management.Automation.PSCredential `
-                                         -ArgumentList @($searchService.ProcessIdentity, $dummyPassword)
-
             $returnVal =  @{
                 Name                        = $serviceApp.DisplayName
                 ProxyName                   = $pName
@@ -143,7 +145,7 @@ function Get-TargetResource
                 SearchCenterUrl             = $serviceApp.SearchCenterUrl
                 DefaultContentAccessAccount = $defaultAccount
                 CloudIndex                  = $cloudIndex
-                WindowsServiceAccount       = $windowsAccount
+                WindowsServiceAccount       = $params.WindowsServiceAccount
                 InstallAccount              = $params.InstallAccount
             }
             return $returnVal
@@ -204,6 +206,12 @@ function Set-TargetResource
     )
 
     Write-Verbose -Message "Setting Search service application '$Name'"
+
+    if ($PSBoundParameters.ContainsKey("WindowsServiceAccount"))
+    {
+        Write-Verbose -Message ("This parameter is deprecated in this resource. Please use " + `
+                                "SPSearchServiceSettings instead.")
+    }
 
     $result = Get-TargetResource @PSBoundParameters
 
@@ -294,15 +302,6 @@ function Set-TargetResource
                     $serviceApp.SearchCenterUrl = $params.SearchCenterUrl
                     $serviceApp.Update()
                 }
-
-                if ($params.ContainsKey("WindowsServiceAccount") -eq $true)
-                {
-                    Write-Verbose -Message ("Setting WindowsServiceAccount to " + `
-                                            $params.WindowsServiceAccount.UserName)
-                    Set-SPEnterpriseSearchService -Identity $params.Name `
-                                                  -ServiceAccount $params.WindowsServiceAccount.UserName `
-                                                  -ServicePassword $params.WindowsServiceAccount.Password
-                }
             }
         }
     }
@@ -387,16 +386,6 @@ function Set-TargetResource
                     }
                 $serviceApp.SearchCenterUrl = $params.SearchCenterUrl
                 $serviceApp.Update()
-            }
-
-            if ($params.ContainsKey("WindowsServiceAccount") -eq $true -and `
-                $result.WindowsServiceAccount.UserName -ne $params.WindowsServiceAccount.UserName)
-            {
-                Write-Verbose -Message ("Updating WindowsServiceAccount to " + `
-                                        $params.WindowsServiceAccount.UserName)
-                Set-SPEnterpriseSearchService -Identity $params.Name `
-                                              -ServiceAccount $params.WindowsServiceAccount.UserName `
-                                              -ServicePassword $params.WindowsServiceAccount.Password
             }
         }
     }
@@ -483,6 +472,12 @@ function Test-TargetResource
 
     Write-Verbose -Message "Testing Search service application '$Name'"
 
+    if ($PSBoundParameters.ContainsKey("WindowsServiceAccount"))
+    {
+        Write-Verbose -Message ("This parameter is deprecated in this resource. Please use " + `
+                                "SPSearchServiceSettings instead.")
+    }
+
     $PSBoundParameters.Ensure = $Ensure
 
     $CurrentValues = Get-TargetResource @PSBoundParameters
@@ -496,20 +491,6 @@ function Test-TargetResource
         if ($desired -ne $current)
         {
             Write-Verbose -Message "Default content access account is different, returning false"
-            Write-Verbose -Message "Desired: $desired. Current: $current."
-            return $false
-        }
-    }
-
-    if ($PSBoundParameters.ContainsKey("WindowsServiceAccount") `
-        -and $Ensure -eq "Present")
-    {
-        $desired = $WindowsServiceAccount.UserName
-        $current = $CurrentValues.WindowsServiceAccount.UserName
-
-        if ($desired -ne $current)
-        {
-            Write-Verbose -Message "Windows service account is different, returning false"
             Write-Verbose -Message "Desired: $desired. Current: $current."
             return $false
         }